% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grunfeld.R
\docType{data}
\name{grunfeld}
\alias{grunfeld}
\title{A data set from Grunfeld (1958)}
\format{
A data set with 200 observations and 6 variables
}
\source{
Grunfeld, Yehuda. 1958. The Determinants of Corporate Investment.
PhD thesis. University of Chicago.
}
\description{
A data set from Grunfeld (1958)
}
\keyword{datasets}
